// src/lib/api-client/models/database-query-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Studio API
 * Studio API for managing campaigns, personas, proxies, and keyword sets
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DatabaseQueryResult
 */
export interface DatabaseQueryResult {
    /**
     * 
     * @type {Array<string>}
     * @memberof DatabaseQueryResult
     */
    'columns'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof DatabaseQueryResult
     */
    'error'?: string;
    /**
     * 
     * @type {number}
     * @memberof DatabaseQueryResult
     */
    'executionTime'?: number;
    /**
     * 
     * @type {number}
     * @memberof DatabaseQueryResult
     */
    'rowCount'?: number;
    /**
     * 
     * @type {Array<Array<object>>}
     * @memberof DatabaseQueryResult
     */
    'rows'?: Array<Array<object>>;
    /**
     * 
     * @type {boolean}
     * @memberof DatabaseQueryResult
     */
    'success'?: boolean;
}